Add setting to make widget transparent
