+++ app/src/main/java/cz/martykan/forecastie/widgets/ExtensiveWidgetProvider.java
@@ -63,6 +63,7 @@ public class ExtensiveWidgetProvider extends AbstractWidgetProvider {
 
             appWidgetManager.updateAppWidget(widgetId, remoteViews);
         }
+        scheduleNextUpdate(context);
     }
 
 }
