+++ app/src/main/res/xml/prefs.xml
@@ -57,6 +57,10 @@
             android:key="theme"
             android:title="@string/setting_theme" />
 
+        <CheckBoxPreference
+            android:defaultValue="false"
+            android:key="transparentWidget"
+            android:title="@string/setting_transparent_widget" />
         <ListPreference
             android:defaultValue="arrow"
             android:entries="@array/windDirectionFormats"
